Fix up guest additions install
Turns out that on first VM boot, system updates might change the kernel version, and if there's not a reboot before installing guest additions, they'll get compiled for the wrong kernel and be unavailable after a subsequent reboot. And it turns out that a subsequent reboot is also necessary in order for Apache to be able to follow symlinks to the shared folder. This was a fun one to track down.